+++ tests/unit/experimental/components/test_graph_pruning.py
@@ -13,27 +13,28 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 from __future__ import annotations
-from typing import Any
-from unittest.mock import patch, Mock, ANY
+
+import datetime
+from typing import Any, Optional
+from unittest.mock import ANY, Mock, patch
 
 import pytest
-
 from neo4j_graphrag.experimental.components.graph_pruning import (
     GraphPruning,
     GraphPruningResult,
     PruningStats,
 )
 from neo4j_graphrag.experimental.components.schema import (
+    GraphSchema,
     NodeType,
     PropertyType,
     RelationshipType,
-    GraphSchema,
 )
 from neo4j_graphrag.experimental.components.types import (
+    LexicalGraphConfig,
+    Neo4jGraph,
     Neo4jNode,
     Neo4jRelationship,
-    Neo4jGraph,
-    LexicalGraphConfig,
 )
 
 
@@ -99,6 +100,44 @@ def test_graph_pruning_filter_properties(
         pruning_stats=PruningStats(),
     )
     assert filtered_properties == expected_filtered_properties
+
+
+@pytest.mark.parametrize(
+    "properties, expected_filtered_properties",
+    [
+        (
+            # all good, no bad types
+            {
+                "name": "John Does",
+                "age": 25,
+                "is_active": True,
+            },
+            {
+                "name": "John Does",
+                "age": 25,
+                "is_active": True,
+            },
+        ),
+        (
+            # map must be serialized
+            {
+                "age": {"dob": datetime.date(2000, 1, 1), "age_in_2025": 25},
+            },
+            {
+                "age": '{"dob": "2000-01-01", "age_in_2025": 25}',
+            },
+        ),
+    ],
+)
+def test_graph_pruning_ensure_property_type(
+    properties: dict[str, Any],
+    expected_filtered_properties: dict[str, Any],
+) -> None:
+    pruner = GraphPruning()
+    type_safe_properties = pruner._ensure_property_types(
+        properties,
+    )
+    assert type_safe_properties == expected_filtered_properties
 
 
 @pytest.fixture(scope="module")
@@ -369,7 +408,7 @@ def test_graph_pruning_validate_relationship(
     additional_relationship_types: bool,
     patterns: tuple[tuple[str, str, str], ...],
     additional_patterns: bool,
-    expected_relationship: str | None,
+    expected_relationship: Optional[str],
     request: pytest.FixtureRequest,
 ) -> None:
     relationship_obj = request.getfixturevalue(relationship)